+++ packages/mongo/collection_async_tests.js
@@ -22,6 +22,7 @@ Tinytest.add('async collection - check for methods presence', function (test) {
 
 ['countDocuments', 'estimatedDocumentCount'].forEach(method => {
   Tinytest.addAsync(`async collection - ${method}`, async test => {
+    if (Meteor.isClient) return;
     const collection = new Mongo.Collection(method + test.id);
     for (let index = 0; index < 10; ++index) {
       test.instanceOf(collection[method](), Promise);